Splits ContainerMetrics and ContainerInfo into two separate structs
Previously GetInfo returned both the container info as well as the metrics.
metrics turned out to be expensive, this change is separating them. It also
removes some unused ContainerInfo fields, e.g. HostIpAddress. Some of those
fields required looping through the network interfaces and figuring out
the ip address. although we could possibly cache the information it seems
unecessary to do the effort given that the containerizer doesn't use them.

[#105015120]
